[feature] 2번째 페이지 작업 - 2번째 페이지 작업
