--- src/App.module.scss
@@ -1,51 +1,0 @@
-@charset "UTF-8";
-
-.app {
-
-  &__wrap {
-
-    padding: 12px;
-    display: flex;
-
-    .left {
-      margin: 4px;
-      flex-grow: 1;
-      flex-basis: 50%;
-    }
-
-    .right {
-      margin: 4px;
-      flex-grow: 1;
-      flex-basis: 50%;
-    }
-  }
-
-  &__user {
-    display: flex;
-    border: 1px solid #74747E;
-
-    align-items: center;
-
-    &:not(:first-child) {
-      margin-top: 10px;
-    }
-
-    .user-info {
-      flex-grow: 1;
-
-      > p {
-        font-size: 12px;
-        letter-spacing: -0.2px;
-        padding: 0 8px;
-      }
-    }
-
-    .pass-button {
-      margin-right: 8px;
-      font-size: 15px;
-      padding: 4px 12px;
-    }
-
-  }
-
-}
