+++ src/components/Index/UserSearch/UserSearch.module.scss
@@ -1,0 +1,18 @@
+@charset "UTF-8";
+
+.user-search {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  > input {
+    padding: 4px 12px;
+    font-size: 15px;
+  }
+
+  > button {
+    font-size: 15px;
+    padding: 4px 12px;
+  }
+
+}
