+++ src/store/atoms.ts
@@ -1,0 +1,4 @@
+import { atom } from 'jotai';
+import { UsersInterface } from '../types/usersTypes';
+
+export const totalPassUsersAtom = atom<UsersInterface[] | []>([]);
